use db instead of memory cache for sso.auth.state

+++ src/modules/rdb/config/yaml.go
@@ -34,7 +34,7 @@ type ssoSection struct {
 	ClientId        string `yaml:"clientId"`
 	ClientSecret    string `yaml:"clientSecret"`
 	ApiKey          string `yaml:"apiKey"`
-	StateExpiresIn  int    `yaml:"stateExpiresIn"`
+	StateExpiresIn  int64  `yaml:"stateExpiresIn"`
 	CoverAttributes bool   `yaml:"coverAttributes"`
 	Attributes      struct {
 		Dispname string `yaml:"dispname"`
